Move URL reading into ExpenseCreateSwitcher, remove standalone component
Instead of a separate component with a fragile fixed delay, integrate
the URL parameter handling directly into ExpenseCreateSwitcher. The
switcher knows exactly when it has loaded the user's dialog preference
and rendered the correct child component. It checks URL params in
OnAfterRenderAsync (after the child is registered) and on LocationChanged
for subsequent navigations.
