+++ src/Money.Blazor.Host/App.razor
@@ -32,7 +32,6 @@
 <AuthenticatedView>
     <ComponetContainer>
         <ExpenseCreateSwitcher />
-        <ExpenseCreateFromUrl />
     </ComponetContainer>
     <ThemeSwitcher />
 </AuthenticatedView>
